refactor: move files around, fix some more issues
