+++ packages/openapi-types/src/schemas/3.1/compliance/contact-object.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { ContactObjectSchema } from './contact-object'
+import { ContactObjectSchema } from '../contact-object'
 
 describe('contact-object', () => {
   describe('ContactObjectSchema', () => {
