+++ packages/openapi-types/src/schemas/3.1/compliance/example-object.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { OperationObjectSchema } from './operation-object'
+import { OperationObjectSchema } from '../operation-object'
 
 describe('example-object', () => {
   describe('ExampleObject', () => {
