+++ packages/openapi-types/src/schemas/3.1/compliance/license-object.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { LicenseObjectSchema } from './license-object'
+import { LicenseObjectSchema } from '../license-object'
 
 describe('license-object', () => {
   describe('LicenseObjectSchema', () => {
