+++ packages/openapi-types/src/schemas/3.1/compliance/path-item-object.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { PathItemObjectSchema } from './path-item-object'
+import { PathItemObjectSchema } from '../path-item-object'
 
 describe('path-item', () => {
   describe('PathItemObjectSchema', () => {
