+++ packages/openapi-types/src/schemas/3.1/compliance/request-body-object.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { RequestBodyObjectSchema } from './request-body-object'
+import { RequestBodyObjectSchema } from '../request-body-object'
 
 describe('request-body-object', () => {
   describe('ContactObjectSchema', () => {
