+++ packages/openapi-types/src/schemas/3.1/compliance/response-object.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { ResponseObjectSchema } from './response-object'
+import { ResponseObjectSchema } from '../response-object'
 
 describe('response-object', () => {
   describe('ContactObjectSchema', () => {
