+++ packages/openapi-types/src/schemas/3.1/compliance/responses-object.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { ResponsesObjectSchema } from './responses-object'
+import { ResponsesObjectSchema } from '../responses-object'
 
 describe('responses-object', () => {
   describe('ResponsesObjectSchema', () => {
