+++ packages/openapi-types/src/schemas/3.1/compliance/server-object.test.ts
@@ -1,6 +1,6 @@
 import { describe, expect, it } from 'vitest'
 
-import { ServerObjectSchema } from './server-object'
+import { ServerObjectSchema } from '../server-object'
 
 describe('server-object', () => {
   describe('ServerObjectSchema', () => {
